Clean up AssemblyAudit attrs and move MemberTableFormatter
- Remove SourceCoverageSummary computed property (duplicates field builder)
- Remove 11 redundant [JsonPropertyName] attrs from AssemblyAudit (SnakeCaseLower policy handles naming)
- Remove [JsonPropertyName("published")] from InspectionResult (same reason)
- Move MemberTableFormatter from Commands/ to Output/ (presentation concern)
